% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_predictors.R
\name{all_predictors}
\alias{all_predictors}
\title{Collate all predictors}
\usage{
all_predictors(path = NULL)
}
\arguments{
\item{path}{\code{character} path to write output.}
}
\value{
\code{SpatRaster}
}
\description{
Collates all predictor variables of insecticide resistance, per Hancock et
al. 2017 \url{https://doi.org/10.1371/journal.pbio.3000633}, specifically
table S9 \url{https://doi.org/10.1371/journal.pbio.3000633.s020} (.docx file)
}
\examples{
\dontrun{
all_predictors()
}
}
