% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irpreds-package.R
\docType{package}
\name{irpreds-package}
\alias{irpreds}
\alias{irpreds-package}
\title{irpreds: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geryan/irpreds}
  \item Report bugs at \url{https://github.com/geryan/irpreds/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
